apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
redis==5.2.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.3

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
